add switcher for install-for-jmx, make jmx tests simpler
